fix: see description
There was approximately a three-day period where I essential forgot about Git. This caused a lot of
code history to be lost. The next few commits are not really separate points in time, but just
different lines of files staged to represent various points in time.
Any commits here may be broken, and I do not guarantee any of them are sensible, or even valid.