test: remove unit testing
This sounds like a bad decision, but for me, it's the most logical move. My brain just doesn't seem
to work on a per-component way, but instead a practical, "does it all work in the end" way. See, if
I were to revamp this entire project, it would be a huge pain to use unit testing, but with e2e,
it's easy as pie. Another issue is constant rerenders. If I want to test if a prop change works, I'd
have to rerender on *every single freaking change*. And e2e testing still works just fine with TDD
whenever I decide to start a feature with a test. Finally, maintaining unit tests *and* e2e tests is
a *lot*. I'm not a team of 50, I'm just one person. I can't do both for *every single component*.
I can only choose one option here, and the option that seems best is e2e. So, au revoir vitest!